Guides in JSON format
Provides a JSON feed of guides, organized by their title and with the
properties "url" and "category". Where category == "Service" we're not
expecting the corresponding guides to be surfaced in a list of research
guides.

Step 1 - create a page with an alias of "guides-json"
Step 2 - ensure these template files are in place

Signed-off-by: Dan Scott <dscott@laurentian.ca>

Signed-off-by: Dan Scott <dscott@laurentian.ca>