Prevent the research guide JSON feed from having dupes
Our SELECT statement will return multiple entries in the event that a given
guide has multiple URL aliases (for example, guides/archeology and
guides/archaeology both pointing to the same node). As we're using PostgreSQL
for our underlying database, we can use the DISTINCT ON (column) syntax to
prevent duplicates from appearing in the feed; and we don't care which URL
alias we get, because hey, they're aliases!

Signed-off-by: Dan Scott <dscott@laurentian.ca>