Move guide menus from top tabs to side column
Currently, navigation within the research guides occurs by clicking
various tabs on the top of the guide.

This has been proven to be problematic, as it clashes with the
navigation between guides that occurs in the menu just above the row of
tabs.

A standard location for navigation on web sites is to use the left-hand
column. If we take advantage of this, it could enable users to build a
mental model that separates the navigation that occurs between different
guides (on the top menu) from the navigation that occurs within the
content on the same guide (the left-hand column).

As an additional benefit, we recover some of the vertical space from the
top of the screen, particularly for guides that have a *lot* of tabs
where the tabs wrap over two or more rows. Finally, it seems (to me) to
be a lot easier to scan the navigation headings, and reduces the clutter
caused by placing similarly-sized tabs vs. the current content heading.

Increasing font size results in the menu moving to the top as a simple
list. This already automatically happens on mobile. And the behaviour
differs between screen resolutions and browser widths. This is not
unexpected; it's part of responsive design to try and provide the best
experience given the constraints of the viewing device.

Once the user starts to zoom the text a decent amount, it makes sense to
stop dedicating the left-hand column to navigation and enable the reader
to use as much of the screen space to view the actual content as
possible.

Signed-off-by: Dan Scott <dscott@laurentian.ca>